+++ atomic/algorithms/__init__.py
@@ -2,3 +2,4 @@
 from atomic.algorithms.pcf import radial_pair_correlation
 from atomic.algorithms.diffusion import einstein_relation
 from atomic.algorithms.slicing import nearest_molecules
+from atomic.algorithms.harmonics import solid_harmonics
