+++ atomic/algorithms/orbitals.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+'''
+Numerical Orbital Functions
+=================================================
+Building discrete molecular orbitals (for visualization) requires a complex set of operations that
+are provided by this module and wrapped into a clean API.
+'''
