+++ atomic/orbital.py
@@ -29,13 +29,9 @@ Orbital information such as centers and energies.
 import re
 import numpy as np
 import pandas as pd
-import sympy as sy
-from sympy.parsing.sympy_parser import parse_expr
-from sympy.physics.secondquant import KroneckerDelta as kr
 from exa import DataFrame, _conf, Series
 from exa.algorithms import meshgrid3d
 from atomic.field import AtomicField
-from atomic.basis import _cartesian_ordering_function, lmap
 
 
 class Orbital(DataFrame):
@@ -84,74 +80,18 @@ class MOMatrix(DataFrame):
         return df.pivot('vector', 'basis_function', 'coefficient').to_sparse().values
 
 
-def solid_harmonics(l, return_all=False, standard_symbols=True):
-    '''
-    Generate a set of spherical solid harmonic functions for a given angular
-    momentum.
-
-        >>> solid_harmonics(0)
-        {(0, 0): 1}
-        >>> solid_harmonics(1, True)
-        {(0, 0): 1, (1, -1): y, (1, 0): z, (1, 1): x}
-
-    These are the real solutions to the :math:`L^{2}` and :math:`L_{z}`
-    operator's eignevalue problems (the complex solutions are called spherical
-    harmonics).
-
-    Args:
-        l (int): Orbital angular moment
-        return_all (bool): If true, return all computed solid harmonics
-        standard_symbols (bool): Convert to standard symbol notation (e.g. x*y => xy)
-
-    Returns:
-        functions (dict): Dictionary of (l, ml) keys and symbolic function values
-    '''
-    x, y, z = sy.symbols('x y z', imaginary=False)
-    r2 = x**2 + y**2 + z**2
-    desired_l = l
-    s = {(0,0): 1}
-    for l in range(1, desired_l + 1):
-        lminus1 = l - 1 if l >= 1 else 0
-        negl = -lminus1 if lminus1 != 0 else 0
-        # top
-        s[(l, l)] = sy.sqrt(2**kr(lminus1, 0) * (2 * lminus1 + 1) / (2 * lminus1 + 2)) * \
-                    (x * s[(lminus1, lminus1)] - (1 - kr(lminus1, 0)) * y * s[(lminus1, negl)])
-        # bottom
-        s[(l, negl - 1)] = sy.sqrt(2**kr(lminus1, 0) * (2 * lminus1 + 1) / (2 * lminus1 + 2)) * \
-                           (y * s[(lminus1, lminus1)] + (1 - kr(lminus1, 0)) * x * s[(lminus1, negl)])
-        for m in range(-l, l + 1)[1:-1]:
-            lminus2 = lminus1 - 1 if lminus1 - 1 >= 0 else 0
-            s_lminus2_m = 0
-            if (lminus2, m) in s:
-                s_lminus2_m = s[(lminus2, m)]
-            s[(l, m)] = ((2 * lminus1 + 1) * z * s[(lminus1, m)] - sy.sqrt((lminus1 + m) * (lminus1 - m)) * \
-                         r2 * s_lminus2_m) / sy.sqrt((lminus1 + m + 1) * (lminus1 - m + 1))
-    # If true, transform the symbolic notation of things like x*y (which represents dxy)
-    # to simply xy (which is also a symbol and therefore possible to manipulate
-    # with .subs({})).
-    if standard_symbols:
-        for i in range(2, desired_l + 1):
-            match0 = r'\*'.join(['([xyz])'] * i)
-            replace0 = r''.join(['\\' + str(j) for j in range(1, i + 1)])
-            match1 = r'([xyz])\*\*(\d+)'
-            for k, v in s.items():
-                if k[0] == i:
-                    expr = re.sub(match0, replace0, str(v.expand()))
-                    for arg, count in re.findall(r'([xyz])\*\*(\d+)', expr):
-                        count = int(count)
-                        f = r''.join([arg[0], ''.join([r'\*'] * count), str(count)])
-                        r = r''.join([arg[0]] * count)
-                        expr = re.sub(f, r, expr)
-                    s[k] = parse_expr(expr)
-    if return_all:
-        return s
-    return {key: value for key, value in s.items() if key[0] == desired_l}
 
 
-def _spherical_gtfs(universe):
+def _voluminate_cartesian_gtfs(universe, xx, yy, zz):
     '''
     Generate symbolic ordered spherical (in cartesian coordinates)
     Gaussian type function basis for the given universe.
+
+    Args:
+        universe: Atomic universe
+        xx (array): Discrete values of x
+        yy (array): Discrete values of y
+        zz (array): Discrete values of z
 
     Note:
         This function exists here because the resultant functions are orbital
@@ -161,7 +101,7 @@ def _spherical_gtfs(universe):
     ordered_spherical_gtf_basis = []
     bases = universe.basis.groupby('symbol')
     lmax = universe.basis['shell'].map(lmap).max()
-    sh = solid_harmonics(lmax, True)
+    sh = solid_harmonics(lmax, True, True)
     for symbol, x, y, z in zip(universe.atom['symbol'], universe.atom['x'],
                                universe.atom['y'], universe.atom['z']):
         bas = bases.get_group(symbol).groupby('function')
@@ -176,6 +116,9 @@ def _spherical_gtfs(universe):
                 function = 0
                 for alpha, c in zip(grp['alpha'], grp['c']):
                     function += c * rx**int(i) * ry**int(j) * rz**int(k) * sy.exp(-alpha * r2)
+                function = sy.lambdify(('x', 'y', 'z'), function, 'numpy')
+                if _conf['pkg_numba']:
+                    from numba import vectorize, float64
                 shell_functions['x' * i + 'y' * j + 'z' * k] = function
             # Reduce the linearly dependent cartesian basis
             # to a linearly independent spherical basis.
